погодозависимость цели

+++ Control/HeatPump.h
@@ -474,9 +474,9 @@ class HeatPump
     uint8_t Chart_Flow_FLOWCON;
     uint8_t Chart_Flow_FLOWEVA;
 
-    float powerOUT;                                         // Мощность выходная, Вт
-    float powerGEO;                                         // Мощность системы GEO, Вт
-    float power220;                                         // Мощность системы 220, Вт
+    int32_t powerOUT;                                       // Мощность выходная, Вт
+    int32_t powerGEO;                                       // Мощность системы GEO, Вт
+    int32_t power220;                                       // Мощность системы 220, Вт
     int32_t fullCOP;                                        // Полный СОР сотые
     int32_t COP;                                            // Чистый COP сотые
